feat(db): fix lng and lat type in UserLocationDbModel

+++ local/db/src/main/java/prus/justweatherapp/local/db/model/UserLocationDbModel.kt
@@ -22,8 +22,8 @@ data class UserLocationDbModel(
     val orderIndex: Int,
 
     @ColumnInfo("lng")
-    val lng: Float,
+    val lng: Double,
 
     @ColumnInfo("lat")
-    val lat: Float,
+    val lat: Double,
 )
